+++ client/src/app/tabs/bpmn/custom/properties-provider/parts/InputOutputProps.js
@@ -10,12 +10,9 @@
 
 import inputOutput from './implementation/InputOutput';
 
-export default function(group, element, bpmnFactory) {
+export default function(group, element, bpmnFactory, translate, options) {
 
-  const inputOutputEntry = inputOutput(element, bpmnFactory);
+  const inputOutputEntry = inputOutput(element, bpmnFactory, translate, options);
+
   group.entries = group.entries.concat(inputOutputEntry.entries);
-  return {
-    getSelectedParameter: inputOutputEntry.getSelectedParameter
-  };
-
 }